01_cosmic_view/requirements.txt
python-binance
numpy
backtrader
flask